requirements.txt
torch_geometric==2.6.0+cpu
torch-scatter==6.0.2
torch-sparse==0.6.18
torch-cluster==1.6.3
torch-geometric==2.6.1
numpy==2.2.3
requests==2.32.3
tqdm==4.67.1
docutils== 0.21.2
sphinx== 8.2.3
PyYAML==6.0.2

